+++ includes/server.hpp
@@ -1,7 +1,7 @@
 #ifndef SERVER_HPP
 #define SERVER_HPP
 
-#include "./../struct.hpp"
+#include "struct.hpp"
 #include "io_multiplexing.hpp"
 #include <iostream>
 #include <vector>
